refactor: domain과 view 분리
domain과 view를 각 폴더에 분리
